+++ benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/ContinuousBenchmark.scala
@@ -23,8 +23,7 @@ object ContinuousBenchmark extends App {
       NearestNeighborsQuery.Exact(field, Similarity.L2),
       Mapping.L2Lsh(randomDenseFloats.dims, 400, 1, 3),
       Seq(
-        Query(NearestNeighborsQuery.L2Lsh(field, 1000), k),
-        Query(NearestNeighborsQuery.L2Lsh(field, 1300, useMLTQuery = true), k)
+        Query(NearestNeighborsQuery.L2Lsh(field, 1000), k)
       )
     ),
     // Angular exact, LSH
@@ -34,8 +33,7 @@ object ContinuousBenchmark extends App {
       NearestNeighborsQuery.Exact(field, Similarity.Angular),
       Mapping.AngularLsh(randomDenseFloats.dims, 400, 1),
       Seq(
-        Query(NearestNeighborsQuery.AngularLsh(field, 1000), k),
-        Query(NearestNeighborsQuery.AngularLsh(field, 1300, useMLTQuery = true), k),
+        Query(NearestNeighborsQuery.AngularLsh(field, 1000), k)
       )
     ),
     // Jaccard exact, sparse indexed, LSH
@@ -45,8 +43,7 @@ object ContinuousBenchmark extends App {
       NearestNeighborsQuery.Exact(field, Similarity.Jaccard),
       Mapping.JaccardLsh(randomSparseBools.dims, 400, 1),
       Seq(
-        Query(NearestNeighborsQuery.JaccardLsh(field, 1000), k),
-        Query(NearestNeighborsQuery.JaccardLsh(field, 1300, useMLTQuery = true), k)
+        Query(NearestNeighborsQuery.JaccardLsh(field, 1000), k)
       )
     )
   )