+++ core/src/main/java/com/klibisz/elastiknn/utils/ArrayUtils.java
@@ -1,5 +1,7 @@
 package com.klibisz.elastiknn.utils;
 
+import java.util.Arrays;
+
 /**
  * Java implementations of some hot spots.
  */
@@ -32,5 +34,66 @@ public class ArrayUtils {
         }
         return n;
     }
+
+    /**
+     * Find the kth largest value in the given array.
+     * Swaps elements in the given array.
+     * Based on: https://github.com/bephrem1/backtobackswe, https://www.youtube.com/watch?v=hGK_5n81drs.
+     * Lucene also has an implementation: https://lucene.apache.org/core/8_0_0/core/org/apache/lucene/util/IntroSelector.html,
+     * but it's more abstract and was slower when I benchmarked it.
+     * @param arr The array.
+     * @param k The position.
+     * @return The index of the kth largest value.
+     */
+    public static int quickSelect(int[] arr, int k) {
+        int n = arr.length;
+        int left = 0;
+        int right = n - 1;
+        int finalIndexOfChoosenPivot = 0;
+        while (left <= right) {
+            int choosenPivotIndex = (right - left + 1) / 2 + left;
+            finalIndexOfChoosenPivot = qsPartition(arr, left, right, choosenPivotIndex);
+            if (finalIndexOfChoosenPivot == n - k) {
+                break;
+            } else if (finalIndexOfChoosenPivot > n - k) {
+                right = finalIndexOfChoosenPivot - 1;
+            } else {
+                left = finalIndexOfChoosenPivot + 1;
+            }
+        }
+        return arr[finalIndexOfChoosenPivot];
+    }
+
+    /**
+     * Same as quickSelect, except makes a copy of the array so the original is unmodified.
+     * @param arr
+     * @param k
+     * @return
+     */
+    public static int quickSelectCopy(int[] arr, int k) {
+        return quickSelect(Arrays.copyOf(arr, arr.length), k);
+    }
+
+
+    private static int qsPartition(int[] arr, int left, int right, int pivotIndex) {
+        int pivotValue = arr[pivotIndex];
+        int lesserItemsTailIndex = left;
+        qsSwap(arr, pivotIndex, right);
+        for (int i = left; i < right; i++) {
+            if (arr[i] < pivotValue) {
+                qsSwap(arr, i, lesserItemsTailIndex);
+                lesserItemsTailIndex++;
+            }
+        }
+        qsSwap(arr, right, lesserItemsTailIndex);
+        return lesserItemsTailIndex;
+    }
+
+    private static void qsSwap(int[] arr, int first, int second) {
+        int temp = arr[first];
+        arr[first] = arr[second];
+        arr[second] = temp;
+    }
+
 
 }