+++ plugin/src/main/scala/com/klibisz/elastiknn/query/KnnQueryBuilder.scala
@@ -8,7 +8,6 @@ import com.google.common.io.BaseEncoding
 import com.klibisz.elastiknn.api.ElasticsearchCodec._
 import com.klibisz.elastiknn.api._
 import com.klibisz.elastiknn.models.{ExactSimilarityFunction, SparseIndexedSimilarityFunction}
-import com.klibisz.elastiknn.storage.StoredVec
 import com.klibisz.elastiknn.utils.CirceUtils.javaMapEncoder
 import com.klibisz.elastiknn.{ELASTIKNN_NAME, api}
 import io.circe.Json
@@ -86,22 +85,22 @@ final case class KnnQueryBuilder(query: NearestNeighborsQuery) extends AbstractQ
         ExactQuery(f, v, ExactSimilarityFunction.Angular)
 
       case (SparseIndexed(f, Similarity.Jaccard, sbv: Vec.SparseBool), _: Mapping.SparseIndexed) =>
-        SparseIndexedQuery(f, sbv, SparseIndexedSimilarityFunction.Jaccard)
+        SparseIndexedQuery(f, sbv, SparseIndexedSimilarityFunction.Jaccard, c.getIndexReader)
 
       case (SparseIndexed(f, Similarity.Hamming, sbv: Vec.SparseBool), _: Mapping.SparseIndexed) =>
-        SparseIndexedQuery(f, sbv, SparseIndexedSimilarityFunction.Hamming)
+        SparseIndexedQuery(f, sbv, SparseIndexedSimilarityFunction.Hamming, c.getIndexReader)
 
-      case (JaccardLsh(f, candidates, v: Vec.SparseBool, useMLTQuery), m: Mapping.JaccardLsh) =>
-        LshQuery(f, m, v, candidates, LshFunctionCache.Jaccard, c.getIndexReader, useMLTQuery)
+      case (JaccardLsh(f, candidates, v: Vec.SparseBool), m: Mapping.JaccardLsh) =>
+        LshQuery(f, v, candidates, LshFunctionCache.Jaccard(m), c.getIndexReader)
 
-      case (HammingLsh(f, candidates, v: Vec.SparseBool, useMLTQuery), m: Mapping.HammingLsh) =>
-        LshQuery(f, m, v, candidates, LshFunctionCache.Hamming, c.getIndexReader, useMLTQuery)
+      case (HammingLsh(f, candidates, v: Vec.SparseBool), m: Mapping.HammingLsh) =>
+        LshQuery(f, v, candidates, LshFunctionCache.Hamming(m), c.getIndexReader)
 
-      case (AngularLsh(f, candidates, v: Vec.DenseFloat, useMLTQuery), m: Mapping.AngularLsh) =>
-        LshQuery(f, m, v, candidates, LshFunctionCache.Angular, c.getIndexReader, useMLTQuery)
+      case (AngularLsh(f, candidates, v: Vec.DenseFloat), m: Mapping.AngularLsh) =>
+        LshQuery(f, v, candidates, LshFunctionCache.Angular(m), c.getIndexReader)
 
-      case (L2Lsh(f, candidates, v: Vec.DenseFloat, useMLTQuery), m: Mapping.L2Lsh) =>
-        LshQuery(f, m, v, candidates, LshFunctionCache.L2, c.getIndexReader, useMLTQuery)
+      case (L2Lsh(f, candidates, v: Vec.DenseFloat), m: Mapping.L2Lsh) =>
+        LshQuery(f, v, candidates, LshFunctionCache.L2(m), c.getIndexReader)
 
       case _ => throw incompatible(mapping, query)
     }